python: Rename dazl.model.writing to dazl.protocols.commands.
