+++ python/dazl/model/__init__.py
@@ -33,4 +33,4 @@ The most important classes:
 
 """
 
-from . import core, ledger, writing
+from . import core, ledger
