+++ python/dazl/protocols/_base.py
@@ -12,8 +12,8 @@ from typing import TYPE_CHECKING, Optional, Sequence, Union
 
 from .. import LOG
 from ..damlast.protocols import SymbolLookup
-from ..model.writing import CommandPayload
 from ..prim import Party
+from ..protocols.commands import CommandPayload
 from ..scheduler import Invoker
 from ..util.typing import safe_cast, safe_optional_cast
 from .events import BaseEvent, ContractFilter, TransactionFilter
